requirements.txt
numpy>=1.26.0,<2.0
chromadb>=0.5.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.23.0
python-multipart>=0.0.9
pillow>=10.0.0
onnxruntime>=1.17.0
